--- manager/tests/api/__init__.py
@@ -1,1 +1,0 @@
-
